Refactor logging crate into modular sinks
