+++ crates/logging/src/sink/mod.rs
@@ -1,0 +1,10 @@
+mod guard;
+mod message_sink;
+mod try_map_writer_error;
+
+pub use guard::LineModeGuard;
+pub use message_sink::MessageSink;
+pub use try_map_writer_error::TryMapWriterError;
+
+#[cfg(test)]
+mod tests;
