+++ tools/line_limits.toml
@@ -48,11 +48,6 @@ max_lines = 1700
 warn_lines = 1650
 
 [[overrides]]
-path = "crates/logging/src/lib.rs"
-max_lines = 1700
-warn_lines = 1650
-
-[[overrides]]
 path = "crates/transport/src/daemon.rs"
 max_lines = 1600
 warn_lines = 1550
